+++ libs/tkwsm/test/conanfile.py
@@ -59,7 +59,7 @@ class test_tkwsmRecipe(ConanFile):
         cmake.install()
 
     def requirements(self):
-        self.requires("tkwsm/0.3.8")
+        self.requires("tkwsm/0.3.9")
         self.requires("tkassert/0.3.4@tket/stable")
         self.requires("tkrng/0.3.3@tket/stable")
-        self.requires("catch2/3.6.0")
+        self.requires("catch2/3.7.0")