+++ pytket/pytket/circuit/__init__.py
@@ -135,7 +135,7 @@ def overload_add_wasm_to_reg(
     if args_wasm is None:
         args_wasm = [0]
 
-    if filehandler._check_file:
+    if filehandler.checked:
         for reg in list_i:
             if reg.size > 32:
                 raise ValueError(
@@ -150,7 +150,11 @@ please use only registers of at most 32 bits"""
 please use only registers of at most 32 bits"""
                 )
 
-    if filehandler.check_function(funcname, len(list_i), len(list_o)):
+    # If the filehandler has not been checked we allow it to
+    # be added without checking the function arity.
+    if not filehandler.checked or filehandler.check_function(
+        funcname, len(list_i), len(list_o)
+    ):
         if (len(args_wasm)) > 0:
             self._add_w_register(max(args_wasm) + 1)
         return self._add_wasm(